desktop/requirements.txt
pynput==1.7.7
pystray==0.19.5
sounddevice==0.5.1
pyperclip==1.9.0
httpx==0.28.1
Pillow==11.2.1
numpy==2.2.5
